+++ Assets/VRM.Samples/Scripts/VRMRuntimeLoader.cs
@@ -2,6 +2,7 @@
 using System;
 using System.IO;
 using System.Runtime.InteropServices;
+using UniGLTF;
 using UnityEngine;
