+++ Assets/VRM10.Samples/Runtime/ViewerUI.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Linq;
+using UniGLTF;
 using UniHumanoid;
 using UnityEngine;
 using UnityEngine.UI;
